requirements.txt
pytest
selenium
jsonpickle
pony
pymysql
mysql-connector-python-rf
lxml-4.6.1-cp38-cp38-win32.whl
allure-pytest


